+++ include/input/idle_manager.h
@@ -4,6 +4,6 @@
 #include <stdint.h>
 
 void idle_manager_heartbeat();
-void idle_manager_task(uint32_t timestamp);
+void idle_manager_task(uint64_t timestamp);
 
 #endif
