+++ include/input/snapback.h
@@ -15,7 +15,7 @@
 #define ARC_MIN_HEIGHT 1850
 
 void snapback_process(analog_data_s *input, analog_data_s *output);
-void snapback_webcapture_task(uint32_t timestamp, analog_data_s *data);
+void snapback_webcapture_task(uint64_t timestamp, analog_data_s *data);
 void snapback_debug_dump_webusb(uint16_t distance, bool crossover);
 
 #endif
